PostgreSQL: Fix disconnect detection, SQLState may be NULL
PostgreSQL won't return an error state if it is not connected to the server, instead it returns NULL.
According to https://www.postgresql.org/docs/current/errcodes-appendix.html
I'd think it should return a "58" or "XX" series code, but it does not.

This also adds 58000 and 58030 SQLStates as disconnect errors in case they ever occur.